chore: anonomize rollbar data
- Moved rollbar initialization to it's own file
- Real change is really the addition of `captureIp: false` and the
transformer that strips away the encryption key from request.url when
sent to rollbar

License: MIT
